V0.7

+++ datetime/src/test/scala/mtest/ArbitaryData.scala
@@ -38,7 +38,7 @@ object ArbitaryData {
     Gen.choose[Long](-dateRange, dateRange).map(d => Date.valueOf(LocalDate.ofEpochDay(d))))
 
   implicit val arbLocalDate: Arbitrary[LocalDate] =
-    Arbitrary(Gen.choose[Long](-dateRange, dateRange).map(d => LocalDate.ofEpochDay(d.toLong)))
+    Arbitrary(Gen.choose[Long](-dateRange, dateRange).map(d => LocalDate.ofEpochDay(d)))
 
   implicit val arbTimestamp: Arbitrary[Timestamp] = Arbitrary(
     genZonedDateTime.map(d => new Timestamp(d.toInstant.getEpochSecond)))
